Fleshed out Gambler class to pass test cases and fixed some other minor mistakes

+++ src/compMethodologyAssignment4/IRandomValueGenerator.java
@@ -4,6 +4,6 @@
 package compMethodologyAssignment4;
 
 public interface IRandomValueGenerator {
-	public double getRandomInt(int min, int max);
+	public int getRandomInt(int min, int max);
 	public boolean getBooleanProbability(double probability);
 }
